requirements.txt
flask
flask_script
python-dotenv
requests
jsonify
flask_mail
